refactor(client): moves useGetFilteredCompanies hook outside forms directory
