+++ src/client/src/content/forms/interface/InterfaceFormBaseFields.tsx
@@ -12,10 +12,10 @@ import { ConditionalWrapper } from "../../../complib/utils";
 import { useGetPurposes } from "../../../data/queries/tyle/queriesPurpose";
 import { useGetRds } from "../../../data/queries/tyle/queriesRds";
 import { useGetTerminals } from "../../../data/queries/tyle/queriesTerminal";
+import { useGetFilteredCompanies } from "../../../hooks/useGetFilteredCompanies";
 import { getValueLabelObjectsFromEnum } from "../../../utils/getValueLabelObjectsFromEnum";
 import { TerminalButton } from "../../common/terminal";
 import { PlainLink } from "../../utils/PlainLink";
-import { useGetFilteredCompanies } from "../common/utils/useGetFilteredCompanies";
 import { resetSubform } from "./InterfaceForm.helpers";
 import { InterfaceFormBaseFieldsContainer } from "./InterfaceFormBaseFields.styled";
 import { InterfaceFormPreview } from "./InterfaceFormPreview";
