+++ src/client/src/content/forms/terminal/TerminalFormBaseFields.tsx
@@ -6,8 +6,8 @@ import { Button } from "../../../complib/buttons";
 import { FormField } from "../../../complib/form";
 import { Input, Select, Textarea } from "../../../complib/inputs";
 import { Flexbox } from "../../../complib/layouts";
+import { useGetFilteredCompanies } from "../../../hooks/useGetFilteredCompanies";
 import { PlainLink } from "../../utils/PlainLink";
-import { useGetFilteredCompanies } from "../common/utils/useGetFilteredCompanies";
 import { TerminalFormBaseFieldsContainer } from "./TerminalFormBaseFields.styled";
 import { TerminalFormPreview } from "./TerminalFormPreview";
 import { FormTerminalLib } from "./types/formTerminalLib";
